Add driver for the riak-java-client
Requires that the java bench driver application is installed
and running also. See the README for that application for
details of how to use.

Added the new rebar as using Jinterface to bridge to Java
so basho_bench app must start in distributed mode, new rebar allows
emu_script_args to set -name and -setcookie.