Add value_generator_blob_file config option to value generator.
If present, e.g.,

    {value_generator_blob_file, "/path/to/a/file"}.

... then the value generator will use the contents of that file instead
of the (mostly) completely random bytes that are generated by
`crypto:rand_bytes/1`.  By allowing the specify an arbitrary file,
the user has control over much entropy is present in the values that
basho_bench generates.  The user can specify a file with a single
repeating byte (which would yield the same values that the basho_bench
value generator `{fixed_char, Size}` will yield, or specify a file
that contains app-specific data such as representative JSON samples,
or a file that contains output from a command like `dd if=/dev/urandom`,
which would yield an extremely low amount of entropy similar to
what the basho_bench key generator `{fixed_bin, Size}` yields.

The main reason for adding this kind of user-controllable entropy
for value blobs is to study the compression rates of Riak storage
backends such as ELevelDB.

Using the ELevelDB backend with compression disabled or enabled
and with various key generators used (and all other basho_bench
configuration knobs identical.

* compression=false
* 1000-of-N ... Using `{fixed_bin, 1000}` and `{value_generator_source_size, N}`
  where N is closer to 1000 -> less entropy -> better possible compression rate
* foo2*.{file-size}KB ... using `{value_generator_blob_file, "file"}`
  where the file size is {file-size} kilobytes, i.e., 100KB or 1.3KB.
  This file is various JSON objects concatenated together into a single
  file.
  Smaller file size -> less entropy -> better possible compression rate

    % du -hs rel/riak/data/leveldb*
    3.1G        rel/riak/data/leveldb.compression=false+1000-of-1048576
    1.7G        rel/riak/data/leveldb.compression=true+1000-of-1020
    1.9G        rel/riak/data/leveldb.compression=true+1000-of-1200
    2.4G        rel/riak/data/leveldb.compression=true+1000-of-2048
    2.6G        rel/riak/data/leveldb.compression=true+1000-of-4096
    1.7G        rel/riak/data/leveldb.compression=true+foo2b.100KB
    1.1G        rel/riak/data/leveldb.compression=true+foo2d.1.3KB