Log abnormal worker exit
Since the abnormal exit is converted to a stop reason of normal it is
never logged by SASL.  The worker could do something like call an
undefined function and the user would be none the wiser.  In fact,
this very thing happened to me causing werid behavior that I couldn't
understand.