Respect shutdown_on_error flag on driver crash
If the driver crashes because of bad code or for some other reason and
the `shutdown_on_error` flag is set then a non-zero exit status should
be returned.  I think the way Basho Bench deals with driver/worker
errors and exit codes is wonky but for this patch I simply fixed the
specific case I need fixed.  Now a driver crash with the
`shutdown_on_error` flag enbled will cause an exit status of 2.