s/erlcassa/erlcql/
erlcassa's thrift implementation clashes with casbench's.
erlcassa is nominally used as a cql client but was set up using thrift
(which casbench does already)
it was clearly labelled as alpha, and remains untouched in 2 years.
erlcql is a real cql client, under active development and does not clash
with casbench