Add a fixed size binary set generator
For testing CRDT sets we often want to generate a set of elements of
elements of a certain size. In the past we used int generators, but
sets of 32bit elements aren't realistic enough. We want to know how
many elements of size X before performance degrades. To that end this
commit adds a generator `{fixed_bin_set, ElemSize, Cardinality}` where
`ElemSize` is the size of a random binary element and `Cardinality` is
the size of the set. It makes use of the existing binary block
generation code in valgen.

It would be nice if we could generate variable size element sets in
future. Maybe just multiple sets from which we pick at random?