Attempt to make listkeys async
Difficult to be sure that behaviour seen in test is as a result of feedback loop if listkeys takes longer (i.e. less requests are sent), rather than the actual cluster reacting to the ongoing listkeys operation.

So now try and make this listkeys request async (but still ensure only one is run at a time)