Fixes bug that prevents the use of custom key and value generators
The function new/2 in both basho_bench_keygen and basho_bench_valgen is
expected to return a function.

But if the generator provided is of type {function, Module, Function,
Args} the new/2 function returns, instead of a function, the key or the
value itself.

This commit fixes this problem by wrapping the erlang:apply call in a
function so that the new/2 function returns a function instead of a
value.