Issues with folsom starting
Running the generated basho_bench escript was failing, due to failures to start applications.   Seems to be expectations that everything in app.src is started when running as a script.

Removed addition of applications from app.src.

Then there was an issue with folsom not starting, as it depended on bear starting (which is a module of functions not an application).  Added an ensure started for bear prior to starting folsom ... and this now at least allows the script to run.